+++ src/core/lib/compression/compression_internal.cc
@@ -156,7 +156,7 @@ CompressionAlgorithmSet::CompressionAlgorithmForLevel(
 
 CompressionAlgorithmSet CompressionAlgorithmSet::FromUint32(uint32_t value) {
   CompressionAlgorithmSet set;
-  for (int i = 0; i < GRPC_COMPRESS_ALGORITHMS_COUNT; i++) {
+  for (size_t i = 0; i < GRPC_COMPRESS_ALGORITHMS_COUNT; i++) {
     if (value & (1u << i)) {
       set.set_.set(i);
     }
@@ -185,7 +185,7 @@ CompressionAlgorithmSet::CompressionAlgorithmSet(
 
 bool CompressionAlgorithmSet::IsSet(
     grpc_compression_algorithm algorithm) const {
-  int i = algorithm;
+  size_t i = static_cast<size_t>(algorithm);
   if (i < GRPC_COMPRESS_ALGORITHMS_COUNT) {
     return set_.is_set(i);
   } else {
@@ -194,7 +194,7 @@ bool CompressionAlgorithmSet::IsSet(
 }
 
 void CompressionAlgorithmSet::Set(grpc_compression_algorithm algorithm) {
-  int i = algorithm;
+  size_t i = static_cast<size_t>(algorithm);
   if (i < GRPC_COMPRESS_ALGORITHMS_COUNT) {
     set_.set(i);
   }
