+++ src/cpp/client/channel_test_peer.cc
@@ -16,8 +16,6 @@
  *
  */
 
-#include <stddef.h>
-
 #include <grpcpp/channel.h>
 #include <grpcpp/test/channel_test_peer.h>
 
@@ -26,7 +24,7 @@
 namespace grpc {
 namespace testing {
 
-size_t ChannelTestPeer::registered_calls() const {
+int ChannelTestPeer::registered_calls() const {
   return grpc_core::Channel::FromC(channel_->c_channel_)
       ->TestOnlyRegisteredCalls();
 }
